backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.6
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.6.0
pydantic-settings==2.1.0
requests==2.31.0
opencv-python-headless==4.9.0.80
ultralytics==8.1.9
torch==2.1.2 --index-url https://download.pytorch.org/whl/cpu
torchvision==0.16.2 --index-url https://download.pytorch.org/whl/cpu
torchaudio==2.1.2 --index-url https://download.pytorch.org/whl/cpu
pytorchvideo @ git+https://github.com/facebookresearch/pytorchvideo.git@28fe037d212663c6a24f373b94cc5d478c8c1e1d
cloudinary==1.38.0
websockets==12.0
python-dotenv==1.0.1
httpx==0.26.0
moviepy==1.0.3
pillow==10.2.0
numpy==1.26.3
imutils==0.5.4
scikit-learn==1.4.0
pandas==2.2.0
matplotlib==3.8.2
seaborn==0.13.2
pytest==8.0.0
pytest-asyncio==0.23.4
black==24.1.1
isort==5.13.2
flake8==7.0.0
mypy==1.8.0
pre-commit==3.6.0
watchfiles==0.21.0
email-validator==2.1.0.post1
itsdangerous==2.1.2
jinja2==3.1.3
markupsafe==2.1.4
orjson==3.9.12
ujson==5.9.0
python-magic==0.4.27
av==11.0.0
decord==0.6.0
fvcore==0.1.5.post20221221
iopath==0.1.10
simplejson==3.19.2
termcolor==2.4.0
yacs==0.1.8
cython
pyyaml
tqdm
scipy
shapely
onnx
onnxruntime
thop
matplotlib-inline
ipython
ipykernel
jupyter
notebook
jupyterlab
jupyter-server
jupyter-client
jupyter-core
nbformat
nbconvert
traitlets
pygments
bleach
defusedxml
entrypoints
pandocfilters
prometheus-client
flower
gunicorn
loguru
tenacity
asyncpg
greenlet
bcrypt
cryptography
pyjwt
httptools
uvloop
hiredis
msgpack
click
click-didyoumean
click-repl
click-plugins
colorama
backoff
backports.zoneinfo;python_version<"3.9"
tzdata;sys_platform == 'win32'
python-dateutil
six
urllib3
idna
certifi
charset-normalizer
soupsieve
beautifulsoup4
lxml
html5lib
webencodings
cssselect
feedparser
pycparser
cffi
asn1crypto
pyopenssl
service_identity
hyperlink
twisted
incremental
constantly
zope.interface
automat
attrs
typing-extensions;python_version<"3.8"
importlib-metadata;python_version<"3.8"
zipp;python_version<"3.8"
contextlib2;python_version<"3.6"
pathlib2;python_version<"3.6"
scandir;python_version<"3.5"
singledispatch;python_version<"3.4"
enum34;python_version<"3.4"
futures;python_version<"3.2"
argparse;python_version<"2.7" or python_version<"3.2"
configparser;python_version<"3.0"
ordereddict;python_version<"2.7"
unittest2;python_version<"2.7"
mock;python_version<"3.3"
funcsigs;python_version<"3.3"
linecache2
traceback2
unittest-xml-reporting
xmlrunner
nose2
nose
pytest-cov
coverage
codecov
